refactor(backend): parse value in service itself
