+++ crates/models/database/src/application_cache.rs
@@ -1,6 +1,6 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.1
 
-use common_models::{ApplicationCacheKey, ApplicationCacheValue};
+use common_models::ApplicationCacheKey;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
@@ -12,7 +12,7 @@ pub struct Model {
     #[sea_orm(column_type = "Json")]
     pub key: ApplicationCacheKey,
     #[sea_orm(column_type = "Json")]
-    pub value: ApplicationCacheValue,
+    pub value: serde_json::Value,
     pub expires_at: Option<DateTimeUtc>,
 }
