+++ crates/models/dependent/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [dependencies]
 async-graphql = { workspace = true }
+chrono = { workspace = true }
 common-models = { path = "../common" }
 config = { path = "../../config" }
 database-models = { path = "../database" }
@@ -14,6 +15,11 @@ media-models = { path = "../media" }
 fitness-models = { path = "../fitness" }
 rust_decimal = { workspace = true }
 schematic = { workspace = true }
+sea-orm = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 serde_with = { workspace = true }
 strum = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["serde_json"]
