+++ crates/providers/src/igdb.rs
@@ -5,12 +5,11 @@ use application_utils::get_base_http_client;
 use async_trait::async_trait;
 use chrono::Datelike;
 use common_models::{
-    ApplicationCacheKey, ApplicationCacheValue, IdObject, NamedObject, PersonSourceSpecifics,
-    SearchDetails, StoredUrl,
+    ApplicationCacheKey, IdObject, NamedObject, PersonSourceSpecifics, SearchDetails, StoredUrl,
 };
 use common_utils::{ryot_log, PAGE_SIZE};
 use database_models::metadata_group::MetadataGroupWithoutId;
-use dependent_models::SearchResults;
+use dependent_models::{ApplicationCacheValue, SearchResults};
 use enums::{MediaLot, MediaSource};
 use itertools::Itertools;
 use media_models::{
