+++ crates/resolvers/statistics/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 [dependencies]
 async-graphql = { workspace = true }
 common-models = { path = "../../models/common" }
+dependent-models = { path = "../../models/dependent" }
 statistics-service = { path = "../../services/statistics" }
 traits = { path = "../../traits" }
