+++ crates/services/cache/Cargo.toml
@@ -10,10 +10,13 @@ common-models = { path = "../../models/common" }
 common-utils = { path = "../../utils/common" }
 config = { path = "../../config" }
 database-models = { path = "../../models/database" }
+dependent-models = { path = "../../models/dependent" }
 sea-orm = { workspace = true }
 sea-query = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
 
 [package.metadata.cargo-machete]
-ignored = ["tracing"]
+ignored = ["serde", "tracing"]
